Move as much models as possible to other apps, see doc for details
`docs/models in applications.md` contains details about issues with
model migrations from `workshops` to other applications. TLDR: it's
difficult.
