+++ amy/extrequests/admin.py
@@ -1,0 +1,12 @@
+from django.contrib import admin
+
+from extrequests.models import (
+    DataAnalysisLevel,
+    DCWorkshopDomain,
+    DCWorkshopTopic,
+)
+
+
+admin.site.register(DataAnalysisLevel)
+admin.site.register(DCWorkshopDomain)
+admin.site.register(DCWorkshopTopic)
